+++ src/mixin.ts
@@ -4,39 +4,46 @@ import {
     ConfigComplex,
     Constructor,
     ConstructorWithStatic,
-    ExtractMixinsProp,
+    ExtractMixinsProp, ExtractMixinsTargets,
     Mixin,
     MixinFull,
     mixinsAfterInit,
-    MixinsProp
+    MixinsProp,
 } from "./types";
 
 ///////////////////////
 
-export function mixin<Name extends string, Config extends AnyObject, PartialKeys extends Array<keyof Config> = []>(
-    config: (Config | ConfigComplex<Name, Config, PartialKeys>) & ThisType<MixinFull<Name, Config>>
-): MixinFull<Name, Config> {
+export function mixin<
+    Name extends string,
+    Config extends AnyObject,
+    PartialKeys extends Array<keyof Config> = [],
+    Target extends AnyObject = AnyObject
+>(
+    config: (Config | ConfigComplex<Name, Config, PartialKeys, Target>) & ThisType<MixinFull<Name, Config, Target>>
+): MixinFull<Name, Config, Target> {
     const cfg: ConfigComplex<Name, Config> = config as any;
-    return {
-        target: undefined as any,
+    const result = {
+        target: undefined as any as Target,
+        origin: {},
         ...cfg,
     };
+    return result as MixinFull<Name, Config, Target>;
 }
 
 /////////////////////////////
 
-export function useMixinsForObject<T extends AnyObject,
-    Mixins extends Array<Mixin<any, any>>,
-    M1 extends Mixin<any, any>,
-    M2 extends Mixin<any, any> = never,
-    M3 extends Mixin<any, any> = never,
-    M4 extends Mixin<any, any> = never,
-    M5 extends Mixin<any, any> = never,
-    M6 extends Mixin<any, any> = never,
-    M7 extends Mixin<any, any> = never,
-    M8 extends Mixin<any, any> = never,
-    M9 extends Mixin<any, any> = never,
-    M10 extends Mixin<any, any> = never,
+export function useMixinsForObject<
+    T extends ExtractMixinsTargets<M1, M2, M3, M4, M5, M6, M7, M8, M9, M10>,
+    M1 extends Mixin<any, any, any>,
+    M2 extends Mixin<any, any, any> = never,
+    M3 extends Mixin<any, any, any> = never,
+    M4 extends Mixin<any, any, any> = never,
+    M5 extends Mixin<any, any, any> = never,
+    M6 extends Mixin<any, any, any> = never,
+    M7 extends Mixin<any, any, any> = never,
+    M8 extends Mixin<any, any, any> = never,
+    M9 extends Mixin<any, any, any> = never,
+    M10 extends Mixin<any, any, any> = never,
     >(
     target: T,
     m1: M1, m2?: M2, m3?: M3, m4?: M4, m5?: M5, m6?: M6, m7?: M7, m8?: M8, m9?: M9, m10?: M10,
@@ -51,29 +58,47 @@ export function useMixinsForObject<T extends AnyObject,
     return target as any;
 }
 
+export function useMixinsForObject_test<
+    T extends ExtractMixinsTargets<M1, M2, M3, M4, M5, M6, M7, M8, M9, M10>,
+    M1 extends Mixin<any, any, any>,
+    M2 extends Mixin<any, any, any> = never,
+    M3 extends Mixin<any, any, any> = never,
+    M4 extends Mixin<any, any, any> = never,
+    M5 extends Mixin<any, any, any> = never,
+    M6 extends Mixin<any, any, any> = never,
+    M7 extends Mixin<any, any, any> = never,
+    M8 extends Mixin<any, any, any> = never,
+    M9 extends Mixin<any, any, any> = never,
+    M10 extends Mixin<any, any, any> = never,
+    >(
+    target: T,
+    m1: M1, m2?: M2, m3?: M3, m4?: M4, m5?: M5, m6?: M6, m7?: M7, m8?: M8, m9?: M9, m10?: M10,
+): T { return target; }
 /////////////////////////////
 
-export function useMixins<Ctor extends Constructor<any>,
+export function useMixins<
+    Target extends ExtractMixinsTargets<M1, M2, M3, M4, M5, M6, M7, M8, M9, M10>,
+    Ctor extends Constructor<Target>,
     Instance extends InstanceType<Ctor>,
     CtorArguments extends ConstructorParameters<Ctor>,
     NewInstance extends Omit<Instance, 'mixins' | '__m_b_type'> &
         Omit<BuildMixinsIntersection<M1, M2, M3, M4, M5, M6, M7, M8, M9, M10>, keyof Instance> &
         { mixins: ExtractMixinsProp<Instance> & MixinsProp<[M1, M2, M3, M4, M5, M6, M7, M8, M9, M10]> },
-    M1 extends Mixin<any, any>,
-    M2 extends Mixin<any, any> = never,
-    M3 extends Mixin<any, any> = never,
-    M4 extends Mixin<any, any> = never,
-    M5 extends Mixin<any, any> = never,
-    M6 extends Mixin<any, any> = never,
-    M7 extends Mixin<any, any> = never,
-    M8 extends Mixin<any, any> = never,
-    M9 extends Mixin<any, any> = never,
-    M10 extends Mixin<any, any> = never,
+    M1 extends Mixin<any, any, any>,
+    M2 extends Mixin<any, any, any> = never,
+    M3 extends Mixin<any, any, any> = never,
+    M4 extends Mixin<any, any, any> = never,
+    M5 extends Mixin<any, any, any> = never,
+    M6 extends Mixin<any, any, any> = never,
+    M7 extends Mixin<any, any, any> = never,
+    M8 extends Mixin<any, any, any> = never,
+    M9 extends Mixin<any, any, any> = never,
+    M10 extends Mixin<any, any, any> = never,
     >(
     klass: Ctor,
     m1: M1, m2?: M2, m3?: M3, m4?: M4, m5?: M5, m6?: M6, m7?: M7, m8?: M8, m9?: M9, m10?: M10,
 ): ConstructorWithStatic<NewInstance, Ctor> & { __m_b_type: Ctor } {
-    return class extends klass {
+    return class extends (klass as any) {
         // tslint:disable-next-line:variable-name
         static __m_b_type = klass;
 
@@ -88,20 +113,24 @@ export function useMixins<Ctor extends Constructor<any>,
     } as any;
 }
 
-export function applyMixins<T extends AnyObject, Mixins extends Array<Mixin<any, any>>>(
+export function applyMixins<T extends AnyObject, Mixins extends Array<Mixin<any, any, any>>>(
     target: T, ...mixins: Mixins
 ): void {
     return applyMixinsInternal(target, false, ...mixins);
 }
 
 function mixingProps<T extends AnyObject, EX extends true | false>(
-    propertyDescriptors: PropertyDescriptorMap, target: T, mixin: Mixin<any, any>
+    propertyDescriptors: PropertyDescriptorMap, target: T, mixin: Mixin<any, any, any>
 ) {
     for (const propName of Object.keys(propertyDescriptors)) {
         const propDescription = propertyDescriptors[propName];
 
         if (!(propName in target)) {
-            if (typeof propDescription.value === 'function') {
+            const isFunction = typeof propDescription.value === 'function';
+            const isGetter = !!propDescription.get;
+            const isSetter = !!propDescription.set;
+
+            if (isFunction) {
                 Reflect.defineProperty(target, propName, {
                     enumerable: false,
                     configurable: true,
@@ -110,17 +139,17 @@ function mixingProps<T extends AnyObject, EX extends true | false>(
                         return mixin[propName](...args);
                     },
                 });
-            } else if (propDescription.get || propDescription.set) {
+            } else if (isGetter || isSetter) {
                 const attributes: PropertyDescriptor = {
                     enumerable: false,
                     configurable: true,
                 };
-                if (propDescription.get) {
+                if (isGetter) {
                     attributes.get = function(this: typeof target) {
                         return mixin[propName];
                     };
                 }
-                if (propDescription.set) {
+                if (isSetter) {
                     attributes.set = function(this: typeof target, value: any) {
                         mixin[propName] = value;
                     };
@@ -142,7 +171,7 @@ function mixingProps<T extends AnyObject, EX extends true | false>(
     }
 }
 
-function applyMixinsInternal<T extends AnyObject, EX extends true | false, Mixins extends Array<Mixin<any, any>>>(
+function applyMixinsInternal<T extends AnyObject, EX extends true | false, Mixins extends Array<Mixin<any, any, any>>>(
     target: T, extendFromMixins: EX, ...mixins: Mixins
 ): void {
     if (typeof target.mixins !== 'object') {
@@ -159,7 +188,7 @@ function applyMixinsInternal<T extends AnyObject, EX extends true | false, Mixin
         mixin.target = target as any;
     }
 
-    for (const mixin of Object.values(target.mixins as Record<string, Mixin<any, any>>)) {
+    for (const mixin of Object.values(target.mixins as Record<string, Mixin<any, any, any>>)) {
         if ('setup' in mixin && typeof mixin.setup === 'function') {
             const setupResult = mixin.setup();
             if (typeof setupResult !== 'undefined') {
@@ -174,9 +203,99 @@ function applyMixinsInternal<T extends AnyObject, EX extends true | false, Mixin
 
             mixingProps(mixables, target, mixin);
         }
+
+        if (typeof mixin.rewrite !== 'undefined') {
+            const rewrites = mixin._rewrite = mixin.rewrite(mixin);
+
+            for (const propName of Object.keys(rewrites)) {
+                const originalDescriptor = getOriginalPropertyDescriptor(target, propName);
+                const rewriteDescriptor = Reflect.getOwnPropertyDescriptor(rewrites, propName);
+                const rewriteItem = rewrites[propName];
+                if (rewriteItem && originalDescriptor && rewriteDescriptor) {
+                    if (!(propName in mixin.origin)) {
+                        Reflect.defineProperty(mixin.origin, propName, originalDescriptor);
+                    }
+
+                    const isRewriteFunction = typeof rewriteDescriptor.value === 'function';
+                    const isRewriteGetter = !!rewriteDescriptor.get;
+                    const isRewriteSetter = !!rewriteDescriptor.set;
+
+                    const isOriginalFunction = typeof originalDescriptor.value === 'function';
+                    const isOriginalGetter = !!originalDescriptor.get;
+                    const isOriginalSetter = !!originalDescriptor.set;
+
+                    const newDescriptor = {...originalDescriptor};
+                    delete newDescriptor.value;
+                    delete newDescriptor.get;
+                    delete newDescriptor.set;
+
+                    if (isOriginalFunction) {
+                        if (isRewriteFunction) {
+                            newDescriptor.value = rewriteDescriptor.value.bind(mixin.target);
+                        } else if (isRewriteGetter) {
+                            newDescriptor.value = function(this: any) {
+                                return rewriteDescriptor.value;
+                            };
+                        }
+                    } else if (isOriginalGetter || isOriginalSetter) {
+                        if (isOriginalGetter) {
+                            if (isRewriteGetter) {
+                                newDescriptor.get = rewriteDescriptor.get;
+                            } else {
+                                newDescriptor.get = originalDescriptor.get;
+                            }
+                        }
+                        if (isOriginalSetter) {
+                            if (isRewriteSetter) {
+                                newDescriptor.set = rewriteDescriptor.set;
+                            } else {
+                                newDescriptor.set = originalDescriptor.set;
+                            }
+                            delete newDescriptor.writable;
+                        }
+                    } else {
+                        if (isRewriteFunction) {
+                            newDescriptor.value = rewriteDescriptor.value();
+                        } else {
+                            // if (isRewriteGetter && !isRewriteSetter && newDescriptor.writable)) {
+                            //     let rVal = originalDescriptor.value;
+                            //
+                            //     newDescriptor.get = function(this: any) {
+                            //         return rVal;
+                            //     };
+                            //     newDescriptor.set = function(this: any, newVal: any) {
+                            //         rVal = newVal;
+                            //     };
+                            //     delete newDescriptor.writable;
+                            if (isRewriteGetter || isRewriteSetter) {
+                                if (isRewriteGetter) {
+                                    newDescriptor.get = rewriteDescriptor.get;
+                                }
+                                if (isRewriteSetter) {
+                                    newDescriptor.set = rewriteDescriptor.set;
+                                }
+                                delete newDescriptor.writable;
+                            } else if (typeof rewriteDescriptor.value !== 'undefined') {
+                                newDescriptor.value = rewriteDescriptor.value;
+                            }
+
+                        }
+                    }
+
+                    if ('get' in newDescriptor || 'set' in newDescriptor || 'value' in newDescriptor) {
+                        Reflect.deleteProperty(target, propName);
+                        try {
+                            Reflect.defineProperty(target, propName, newDescriptor);
+                        } catch(e) {
+                            debugger
+                        }
+                    }
+                }
+            }
+        }
     }
 
-    for (const mixin of Object.values(target.mixins as Record<string, Mixin<any, any>>)) {
+    for (const mixin of Object.values(target.mixins as Record<string, Mixin<any, any, any>>)) {
         mixin.init && mixin.init();
     }
 
@@ -207,4 +326,14 @@ function getMixables(obj: object): PropertyDescriptorMap {
     }
 
     return map;
+}
+
+export function getOriginalPropertyDescriptor(object: {}, key: string): PropertyDescriptor | undefined {
+    if (object.hasOwnProperty(key)) {
+        return Object.getOwnPropertyDescriptor(object, key);
+    } else if (key in object) {
+        return getOriginalPropertyDescriptor(Object.getPrototypeOf(object), key);
+    } else {
+        return undefined;
+    }
 }